Restructure the frontend.
