+++ frontend/src/components/Authenticated/Logout.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from "react-router-dom";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../Context/AuthContext";
 import { useEffect } from "react";
 
 function Logout() {
