+++ frontend/src/pages/NotAuthenticated/LoginPage.tsx
@@ -1,4 +1,4 @@
-import LoginForm from "../components/LoginForm";
+import LoginForm from "../../components/NotAuthenticated/LoginForm";
 
 function LoginPage() {
   return <LoginForm />;
