+++ frontend/src/pages/index.tsx
@@ -1,4 +1,4 @@
 export { default as HomePage } from "./Home/HomePage";
-export { default as LoginPage } from "./LoginPage";
-export { default as RegisterPage } from "./RegisterPage";
+export { default as LoginPage } from "./NotAuthenticated/LoginPage";
+export { default as RegisterPage } from "./NotAuthenticated/RegisterPage";
 export { default as TodoListsPage } from "./Authenticated/TodoListsPage";
